тестовый релиз

+++ BadMom.DAL/Interfaces/IRepository.cs
@@ -9,10 +9,10 @@ namespace BadMom.DAL.Interfaces
     public interface IRepository<T> where T : class
     {
         IEnumerable<T> GetAll();
-        T Get(int id);
+        T Get(long id);
         IEnumerable<T> Find(Func<T, Boolean> predicate);
         void Create(T item);
         void Update(T item);
-        void Delete(int id);
+        void Delete(long id);
     }
 }
